+++ components/context/ProductLandingContext.tsx
@@ -5,6 +5,10 @@ export type TocItem = {
   fullPath: string
   title: string
   intro?: string
+  childTocItems?: Array<{
+    fullPath: string
+    title: string
+  }>
 }
 export type FeaturedLink = {
   title: string
@@ -13,6 +17,7 @@ export type FeaturedLink = {
   authors?: Array<string>
   hideIntro?: boolean
   date?: string
+  fullTitle?: string
 }
 export type CodeExample = {
   title: string
@@ -39,7 +44,7 @@ export type ProductLandingContextT = {
     overview?: string
   } | null
   product_video?: string
-  guideCards: Array<FeaturedLink>
+  featuredLinks: Record<string, Array<FeaturedLink>>
   productCodeExamples: Array<CodeExample>
   productUserExamples: Array<{ username: string; description: string }>
   productCommunityExamples: Array<{ repo: string; description: string }>
@@ -48,10 +53,16 @@ export type ProductLandingContextT = {
     viewAllHref?: string // If provided, adds a "View All ->" to the header
     articles: Array<FeaturedLink>
   }>
-  changelog: { label: string; prefix: string }
   changelogUrl?: string
   whatsNewChangelog?: Array<{ href: string; title: string; date: string }>
   tocItems: Array<TocItem>
+  hasGuidesPage: boolean
+  releases: Array<{
+    version: string
+    firstPreviousRelease: string
+    secondPreviousRelease: string
+    patches: Array<{ date: string; version: string }>
+  }>
 }
 
 export const ProductLandingContext = createContext<ProductLandingContextT | null>(null)
@@ -68,9 +79,27 @@ export const useProductLandingContext = (): ProductLandingContextT => {
   return context
 }
 
+export const getFeaturedLinksFromReq = (req: any): Record<string, Array<FeaturedLink>> => {
+  return Object.fromEntries(
+    Object.entries(req.context.featuredLinks || {}).map(([key, entries]) => {
+      return [
+        key,
+        ((entries as Array<any>) || []).map((entry: any) => ({
+          href: entry.href,
+          title: entry.title,
+          intro: entry.intro,
+          authors: entry.page.authors || [],
+          fullTitle: entry.fullTitle,
+        })),
+      ]
+    })
+  )
+}
+
 export const getProductLandingContextFromRequest = (req: any): ProductLandingContextT => {
   const productTree = req.context.currentProductTree
   const page = req.context.page
+  const hasGuidesPage = (page.children || []).includes('/guides')
   return {
     ...pick(page, [
       'title',
@@ -79,8 +108,8 @@ export const getProductLandingContextFromRequest = (req: any): ProductLandingCon
       'beta_product',
       'intro',
       'product_video',
-      'changelog',
     ]),
+    hasGuidesPage,
     product: {
       href: productTree.href,
       title: productTree.renderedShortTitle || productTree.renderedFullTitle,
@@ -89,6 +118,7 @@ export const getProductLandingContextFromRequest = (req: any): ProductLandingCon
     changelogUrl: req.context.changelogUrl || [],
     productCodeExamples: req.context.productCodeExamples || [],
     productCommunityExamples: req.context.productCommunityExamples || [],
+    releases: req.context.releases || [],
 
     productUserExamples: (req.context.productUserExamples || []).map(
       ({ user, description }: any) => ({
@@ -105,16 +135,7 @@ export const getProductLandingContextFromRequest = (req: any): ProductLandingCon
         }
       : null,
 
-    guideCards: (req.context.featuredLinks ? req.context.featuredLinks.guideCards || [] : []).map(
-      (link: any) => {
-        return {
-          href: link.href,
-          title: link.title,
-          intro: link.intro,
-          authors: link.page.authors || [],
-        }
-      }
-    ),
+    featuredLinks: getFeaturedLinksFromReq(req),
 
     tocItems: req.context.tocItems || [],
 
@@ -124,10 +145,13 @@ export const getProductLandingContextFromRequest = (req: any): ProductLandingCon
       })
       .map(([key, links]: any) => {
         return {
-          label: req.context.site.data.ui.toc[key],
+          label:
+            key === 'popular'
+              ? req.context.page.featuredLinks.popularHeading || req.context.site.data.ui.toc[key]
+              : req.context.site.data.ui.toc[key],
           viewAllHref:
-            key === 'guides' && !req.context.currentCategory
-              ? `${req.context.currentPath}/${key}`
+            key === 'guides' && !req.context.currentCategory && hasGuidesPage
+              ? `${req.context.currentPath}/guides`
               : '',
           articles: links.map((link: any) => {
             return {
@@ -136,6 +160,7 @@ export const getProductLandingContextFromRequest = (req: any): ProductLandingCon
               title: link.title,
               intro: link.intro,
               authors: link.page.authors || [],
+              fullTitle: link.fullTitle,
             }
           }),
         }