+++ components/context/TocLandingContext.tsx
@@ -1,4 +1,6 @@
+import pick from 'lodash/pick'
 import { createContext, useContext } from 'react'
+import { FeaturedLink, getFeaturedLinksFromReq } from './ProductLandingContext'
 
 export type TocItem = {
   fullPath: string
@@ -9,7 +11,11 @@ export type TocItem = {
 export type TocLandingContextT = {
   title: string
   introPlainText: string
+  productCallout: string
   tocItems: Array<TocItem>
+  variant?: 'compact' | 'expanded'
+  featuredLinks: Record<string, Array<FeaturedLink>>
+  renderedPage: string
 }
 
 export const TocLandingContext = createContext<TocLandingContextT | null>(null)
@@ -25,9 +31,17 @@ export const useTocLandingContext = (): TocLandingContextT => {
 }
 
 export const getTocLandingContextFromRequest = (req: any): TocLandingContextT => {
+  const isEarlyAccess = req.context.page?.documentType === 'early-access'
   return {
-    title: req.context.page.title,
+    title: req.context.page.titlePlainText,
+    productCallout: req.context.page.product || '',
     introPlainText: req.context.page.introPlainText,
-    tocItems: req.context.tocItems || [],
+    tocItems: (req.context.genericTocFlat || req.context.genericTocNested || []).map((obj: any) =>
+      pick(obj, ['fullPath', 'title', 'intro', 'childTocItems'])
+    ),
+    variant: req.context.genericTocFlat ? 'expanded' : 'compact',
+
+    featuredLinks: getFeaturedLinksFromReq(req),
+    renderedPage: isEarlyAccess ? req.context.renderedPage : '',
   }
 }