+++ pages/[versionId]/index.tsx
@@ -1,0 +1,4 @@
+import LandingPage from '../index'
+export { getServerSideProps } from '../index'
+
+export default LandingPage