+++ pages/_app.tsx
@@ -3,14 +3,25 @@ import App from 'next/app'
 import type { AppProps, AppContext } from 'next/app'
 import Head from 'next/head'
 import { useTheme, ThemeProvider } from '@primer/components'
-import { getThemeProps } from 'components/lib/getThemeProps'
+import { defaultComponentThemeProps, getThemeProps } from 'components/lib/getThemeProps'
 
 import '../stylesheets/index.scss'
 
-import { defaultThemeProps } from 'components/lib/getThemeProps'
+import events from 'components/lib/events'
+import experiment from 'components/lib/experiment'
+import { LanguagesContext, LanguagesContextT } from 'components/context/LanguagesContext'
 
-type MyAppProps = AppProps & { csrfToken: string; themeProps: typeof defaultThemeProps }
-const MyApp = ({ Component, pageProps, csrfToken, themeProps }: MyAppProps) => {
+type MyAppProps = AppProps & {
+  csrfToken: string
+  themeProps: typeof defaultComponentThemeProps
+  languagesContext: LanguagesContextT
+}
+const MyApp = ({ Component, pageProps, csrfToken, themeProps, languagesContext }: MyAppProps) => {
+  useEffect(() => {
+    events()
+    experiment()
+  }, [])
+
   return (
     <>
       <Head>
@@ -32,9 +43,11 @@ const MyApp = ({ Component, pageProps, csrfToken, themeProps }: MyAppProps) => {
 
         <meta name="csrf-token" content={csrfToken} />
       </Head>
-      <ThemeProvider>
-        <SetTheme themeProps={themeProps} />
-        <Component {...pageProps} />
+      <ThemeProvider dayScheme={themeProps.dayTheme} nightScheme={themeProps.nightTheme}>
+        <LanguagesContext.Provider value={languagesContext}>
+          <SetTheme themeProps={themeProps} />
+          <Component {...pageProps} />
+        </LanguagesContext.Provider>
       </ThemeProvider>
     </>
   )
@@ -46,10 +59,15 @@ MyApp.getInitialProps = async (appContext: AppContext) => {
   const appProps = await App.getInitialProps(appContext)
   const req: any = ctx.req
 
-  return { ...appProps, themeProps: getThemeProps(req), csrfToken: req?.csrfToken?.() || '' }
+  return {
+    ...appProps,
+    themeProps: getThemeProps(req),
+    csrfToken: req?.csrfToken?.() || '',
+    languagesContext: { languages: req.context.languages },
+  }
 }
 
-const SetTheme = ({ themeProps }: { themeProps: typeof defaultThemeProps }) => {
+const SetTheme = ({ themeProps }: { themeProps: typeof defaultComponentThemeProps }) => {
   // Cause primer/components to re-evaluate the 'auto' color mode on client side render
   const { setColorMode } = useTheme()
   useEffect(() => {