+++ script/sync-search-indices.js
@@ -7,16 +7,17 @@
 //
 // [end-readme]
 
-require('make-promises-safe')
+import searchSync from './search/sync.js'
+import 'make-promises-safe'
 
 main()
 
-async function main () {
-  const sync = require('./search/sync')
+async function main() {
+  const sync = searchSync
   const opts = {
     dryRun: 'DRY_RUN' in process.env,
     language: process.env.LANGUAGE,
-    version: process.env.VERSION
+    version: process.env.VERSION,
   }
   await sync(opts)
 }