+++ translations/ja-JP/content/admin/release-notes.md
@@ -1,11 +1,13 @@
 ---
 title: リリースノート
-intro: '{{ allVersions[currentVersion].versionTitle }} のリリースノートです。'
 layout: release-notes
 versions:
   enterprise-server: '*'
+  github-ae: '*'
 topics:
   - Enterprise
   - Upgrades
+redirect_from:
+  - /admin/overview/github-ae-release-notes
 ---